move signing interface to mnemonic node impl
