+++ packages/jellyfish-wallet-mnemonic/src/mnemonic.ts
@@ -1,11 +1,12 @@
-import { SigningInterface, WalletHdNode, WalletHdNodeProvider } from '@defichain/jellyfish-wallet'
+import { SmartBuffer } from 'smart-buffer'
+import { WalletHdNode, WalletHdNodeProvider } from '@defichain/jellyfish-wallet'
 import { DERSignature } from '@defichain/jellyfish-crypto'
 import {
   Transaction,
   Vout,
   TransactionSegWit,
   TransactionSigner,
-  SignInputOption, SIGHASH
+  SignInputOption, SIGHASH, CTransaction
 } from '@defichain/jellyfish-transaction'
 import * as bip32 from 'bip32'
 import * as bip39 from 'bip39'
@@ -79,10 +80,12 @@ export function mnemonicToSeed (mnemonic: string[]): Buffer {
 export class MnemonicHdNode implements WalletHdNode {
   private readonly root: bip32.BIP32Interface
   private readonly path: string
+  private readonly signingCb?: SigningInterface
 
-  constructor (root: bip32.BIP32Interface, path: string) {
+  constructor (root: bip32.BIP32Interface, path: string, signingCb?: SigningInterface) {
     this.root = root
     this.path = path
+    this.signingCb = signingCb
   }
 
   /**
@@ -124,9 +127,22 @@ export class MnemonicHdNode implements WalletHdNode {
     const inputs: SignInputOption[] = prevouts.map(prevout => {
       return { prevout: prevout, ellipticPair: this }
     })
-    return TransactionSigner.sign(transaction, inputs, {
+
+    const signed = await TransactionSigner.sign(transaction, inputs, {
       sigHashType: SIGHASH.ALL
     })
+
+    if (this.signingCb !== undefined) {
+      const unsignedBuffer = new SmartBuffer()
+      new CTransaction(transaction).toBuffer(unsignedBuffer)
+      await this.signingCb.unsigned(unsignedBuffer.toBuffer(), transaction)
+
+      const signedBuffer = new SmartBuffer()
+      new CTransaction(signed).toBuffer(signedBuffer)
+      await this.signingCb.signed(signedBuffer.toBuffer(), signed)
+    }
+
+    return signed
   }
 
   /**
@@ -151,27 +167,35 @@ export class MnemonicHdNode implements WalletHdNode {
   }
 }
 
+export interface SigningInterface {
+  unsigned: (buffer: Buffer, tx: Transaction) => Promise<void>
+  signed: (buffer: Buffer, tx: TransactionSegWit) => Promise<void>
+}
+
 /**
  * Provider that derive MnemonicHdNode from root. Uses a lite on demand derivation.
  */
-export class MnemonicHdNodeProvider extends WalletHdNodeProvider<MnemonicHdNode> {
+export class MnemonicHdNodeProvider implements WalletHdNodeProvider<MnemonicHdNode> {
+  walletHdNode?: MnemonicHdNode
+  signingCb?: SigningInterface
+
   /**
    * @param {Buffer} seed of the hd node
    * @param {Bip32Options} options for chain agnostic generation of public/private keys
    */
-  static fromSeed (seed: Buffer, options: Bip32Options): MnemonicHdNodeProvider {
+  static fromSeed (seed: Buffer, options: Bip32Options, signingCb?: SigningInterface): MnemonicHdNodeProvider {
     const node = bip32.fromSeed(seed, options)
-    return new MnemonicHdNodeProvider(node)
+    return new MnemonicHdNodeProvider(node, signingCb)
   }
 
   private readonly root: bip32.BIP32Interface
 
   private constructor (root: bip32.BIP32Interface, signingCb?: SigningInterface) {
-    super(signingCb)
     this.root = root
+    this.signingCb = signingCb
   }
 
   derive (path: string): MnemonicHdNode {
-    return new MnemonicHdNode(this.root, path)
+    return new MnemonicHdNode(this.root, path, this.signingCb)
   }
 }
